@@ -1,9 +1,5 @@
 class IndexController < ApplicationController
+  # get /
   def index
-    @photos = Photo.recent_photos
-    @posts  = Post.recent.first(3)
-  end
-
-  def about
   end
 end
