@@ -1,49 +1,44 @@
-require 'mina/bundler'
 require 'mina/rails'
 require 'mina/git'
-require 'mina/rbenv'  # for rbenv support. (http://rbenv.org)
+require 'mina/rbenv'  # for rbenv support. (https://rbenv.org)
 
-set :domain, 'maxim.photo'
+# Basic settings:
+#   domain       - The hostname to SSH to.
+#   deploy_to    - Path to deploy into.
+#   repository   - Git repo to clone from. (needed by mina/git)
+#   branch       - Branch name to deploy. (needed by mina/git)
+
+set :application_name, 'maxim_photo'
+set :domain, '89.108.104.88'
 set :deploy_to, '/var/www/maxim.photo'
 set :repository, 'git@github.com:ozgg/maxim.photo.git'
 set :branch, 'master'
 
-# Manually create these paths in shared/ (eg: shared/config/database.yml) in your server.
-# They will be linked in the 'deploy:link_shared_paths' step.
-set :shared_paths, %w(config/database.yml config/secrets.yml log tmp public/uploads)
-
 # Optional settings:
-#   set :user, 'foobar'    # Username in the server to SSH to.
-#   set :port, '30000'     # SSH port number.
+set :user, 'developer'          # Username in the server to SSH to.
+#   set :port, '30000'           # SSH port number.
 #   set :forward_agent, true     # SSH forward_agent.
 
-# This task is the environment that is loaded for most commands, such as
+# shared dirs and files will be symlinked into the app-folder by the 'deploy:link_shared_paths' step.
+set :shared_dirs, fetch(:shared_dirs, []).push('log', 'tmp', 'public/uploads')
+set :shared_files, fetch(:shared_files, []).push('.env')
+
+# This task is the environment that is loaded for all remote run commands, such as
 # `mina deploy` or `mina rake`.
 task :environment do
   invoke :'rbenv:load'
 end
 
-# Put any custom mkdir's in here for when `mina setup` is ran.
-# For Rails apps, we'll make some of the shared paths that are shared between
-# all releases.
-task :setup => :environment do
-  queue! %[mkdir -p "#{deploy_to}/shared/log"]
-  queue! %[mkdir -p "#{deploy_to}/shared/config"]
-  queue! %[mkdir -p "#{deploy_to}/shared/public/uploads"]
-  queue! %[mkdir -p "#{deploy_to}/shared/tmp/pids"]
-
-  queue! %[chmod -R g+rx,u+rwx "#{deploy_to}/shared"]
-
-  queue! %[touch "#{deploy_to}/shared/config/database.yml"]
-  queue! %[touch "#{deploy_to}/shared/config/secrets.yml"]
-  queue  %[echo "-----> Be sure to edit 'shared/config/database.yml' and 'secrets.yml'."]
+# Put any custom commands you need to run at setup
+# All paths in `shared_dirs` and `shared_paths` will be created on their own.
+task :setup do
+  # command %{rbenv install 2.3.0}
 end
 
 desc "Deploys the current version to the server."
-task :deploy => :environment do
-  to :before_hook do
-    # Put things to run locally before ssh
-  end
+task :deploy do
+  # uncomment this line to make sure you pushed your local branch to the remote origin
+  # invoke :'git:ensure_pushed'
   deploy do
     # Put things that will set up an empty directory into a fully set-up
     # instance of your project.
@@ -53,10 +48,19 @@ task :deploy => :environment do
     invoke :'rails:db_migrate'
     invoke :'rails:assets_precompile'
     invoke :'deploy:cleanup'
-    queue! "#{rake} sitemap:generate"
 
-    to :launch do
-      queue! "bundle exec pumactl -S #{deploy_to}/shared/tmp/puma.state restart"
+    on :launch do
+      in_path(fetch(:current_path)) do
+        command %{mkdir -p tmp/}
+        command %{touch tmp/restart.txt}
+      end
     end
   end
+
+  # you can use `run :local` to run tasks on local machine before of after the deploy scripts
+  # run(:local){ say 'done' }
 end
+
+# For help in making your deploy script, see the Mina documentation:
+#
+#  - https://github.com/mina-deploy/mina/tree/master/docs
