@@ -1,2 +1,2 @@
-require "i18n/backend/pluralization"
+require 'i18n/backend/pluralization'
 I18n::Backend::Simple.send(:include, I18n::Backend::Pluralization)
