@@ -1,15 +1,59 @@
-require 'pathname'
+# Puma can serve each request in a thread from an internal thread pool.
+# The `threads` method setting takes two numbers a minimum and maximum.
+# Any libraries that use thread pools should be configured to match
+# the maximum value specified for Puma. Default is set to 5 threads for minimum
+# and maximum, this matches the default thread size of Active Record.
+#
+threads_count = ENV.fetch("RAILS_MAX_THREADS") { 5 }.to_i
+threads threads_count, threads_count
 
-environment 'production'
+# Specifies the `port` that Puma will listen on to receive requests, default is 3000.
+#
+# port        ENV.fetch("PORT") { 3000 }
 
-app_root = Pathname.new('../../../../current').expand_path(__FILE__)
-logs_dir = app_root.join('log')
-tmp_dir  = app_root.join('tmp')
-pids_dir = tmp_dir.join('pids')
+# Specifies the `environment` that Puma will run in.
+#
+environment ENV.fetch("RAILS_ENV") { "development" }
 
-pidfile pids_dir.join('puma.pid').to_s
-state_path tmp_dir.join('puma.state').to_s
+# Specifies the number of `workers` to boot in clustered mode.
+# Workers are forked webserver processes. If using threads and workers together
+# the concurrency of the application would be max `threads` * `workers`.
+# Workers do not work on JRuby or Windows (both of which do not support
+# processes).
+#
+# workers ENV.fetch("WEB_CONCURRENCY") { 2 }
 
-bind "unix://#{tmp_dir}/puma.sock"
+# Use the `preload_app!` method when specifying a `workers` number.
+# This directive tells Puma to first boot the application and load code
+# before forking the application. This takes advantage of Copy On Write
+# process behavior so workers use less memory. If you use this option
+# you need to make sure to reconnect any threads in the `on_worker_boot`
+# block.
+#
+# preload_app!
 
-stdout_redirect logs_dir.join('stdout.log').to_s, logs_dir.join('stderr.log').to_s, true
+# The code in the `on_worker_boot` will be called if you are using
+# clustered mode by specifying a number of `workers`. After each worker
+# process is booted this block will be run, if you are using `preload_app!`
+# option you will want to use this block to reconnect to any threads
+# or connections that may have been created at application boot, Ruby
+# cannot share connections between processes.
+#
+# on_worker_boot do
+#   ActiveRecord::Base.establish_connection if defined?(ActiveRecord)
+# end
+
+if ENV['RAILS_ENV'] == 'production'
+  shared_path = '/var/www/maxim.photo/shared'
+  logs_dir    = "#{shared_path}/log"
+
+  state_path "#{shared_path}/tmp/puma/state"
+  pidfile "#{shared_path}/tmp/puma/pid"
+  bind "unix://#{shared_path}/tmp/puma.sock"
+  stdout_redirect "#{logs_dir}/stdout.log", "#{logs_dir}/stderr.log", true
+
+  activate_control_app
+end
+
+# Allow puma to be restarted by `rails restart` command.
+plugin :tmp_restart
